style(api): cleanup promo routes
